splits forward/inverse, and base HDF5 vs yaml
following Dimitri's suggestion, I included 3 parameter yaml's: a base one, which is the original microzones_complete_nest.yaml we received from pavia; one for the forward block; one for the inverse block. As you can see, both of these import the microzones_complete_nest and only overwrite the parameters that they want to change. i think this is a sensible way to proceed

in addition, by manually setting the simulation in the scaffold, i'm sure that no parameters are written to the base HDF5, keeping it pristine.
related to #14